일단 구조 수정
