--- src/interfaces/user.ts
@@ -1,6 +1,0 @@
-export interface User {
-  email: string
-  name: string
-  position: string
-  password: string
-}
