+++ tailwind.config.ts
@@ -1,8 +1,11 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: ['./src/**/*.{js,ts,jsx,tsx}'],
+// tailwind.config.ts
+import type {Config} from 'tailwindcss'
+
+const config: Config = {
+  content: ['./src/pages/**/*.{js,ts,jsx,tsx}', './src/components/**/*.{js,ts,jsx,tsx}'],
   theme: {
     extend: {}
   },
   plugins: []
 }
+export default config
