--- utils/sample-data.ts
@@ -1,9 +1,0 @@
-import { User } from "../src/interfaces";
-
-/** Dummy user data. */
-export const sampleUserData: User[] = [
-  { id: 101, name: "Alice" },
-  { id: 102, name: "Bob" },
-  { id: 103, name: "Caroline" },
-  { id: 104, name: "Dave" },
-];
